// src/types/project-edge.ts
import { Edge } from '@antv/x6';

interface IEdgeType {
  id: string;
  name: string;
  color: string;
}

export type ProjectEdgeResponse = {
  id?: string;
  name: string;
  target_id: string;
  source_id: string;
  source_attribute_id?: string;
  inverse: boolean;
  multiple: boolean;
  target?: IEdgeType;
  source?: IEdgeType;
  edgeTypes?: { name: string };
  project_edge_type_id?: string;
  template_id?: string;
} & { project_id?: string };

export type ProjectEdgeForm = {
  name: string;
  source_id: string;
  target_id: string;
  inverse: boolean;
  multiple: boolean;
};

export type IOpenId = {
  id: string;
};

export type IOpenEdge = {
  edge: Edge;
};